ISSUE #3296 - set form padding via styled component
