+++ frontend/src/v5/ui/routes/dashboard/projects/federations/editFederationModal/editFederationModal.component.tsx
@@ -22,9 +22,10 @@ import { IFederation } from '@/v5/store/federations/federations.types';
 
 import { FederationsActionsDispatchers } from '@/v5/services/actionsDispatchers/federationsActions.dispatchers';
 import { useParams } from 'react-router-dom';
-import { FormModal, IFormModal } from '@controls/modal/formModal/formDialog.component';
+import { IFormModal } from '@controls/modal/formModal/formDialog.component';
 import { DashboardParams } from '@/v5/ui/routes/routes.constants';
 import { EditFederation } from './editFederation';
+import { FormModal } from './editFederationModal.styles';
 import { useContainersData } from '../../containers/containers.hooks';
 
 type EditFederationModalProps = IFormModal & {
@@ -72,7 +73,6 @@ export const EditFederationModal = ({
 			onSubmit={saveChanges}
 			isValid={includedContainers.length > 0}
 			maxWidth="lg"
-			zeroMargin
 			{...otherProps}
 		>
 			<EditFederation
